fix: change user

+++ backend/resolvers/Mutation/usuario.js
@@ -75,7 +75,7 @@ const mutations = {
             const usuario = await obterUsuario(_, { filtro })
             if(usuario) {
                 const { id } = usuario
-                if(dados.perfis) {
+                if(ctx.admin && dados.perfis) {
                     await db("usuarios_perfis")
                         .where({ usuario_id: id }).delete()
